Refactoring namespace
